backend/requirements.txt
Django==4.2.30
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-django==4.8.0
pytest-cov==4.1.0
